tech(api): move controller to bounded context
