+++ api/tests/unit/application/users/user-controller_test.js
@@ -203,40 +203,6 @@ describe('Unit | Controller | user-controller', function () {
     });
   });
 
-  describe('#getUserCampaignParticipationToCampaign', function () {
-    it('should return serialized campaign participation', async function () {
-      // given
-      const userId = 789;
-      const campaignId = 456;
-      const campaignParticipation = Symbol('campaign participation');
-      const expectedCampaignParticipation = Symbol('expected campaign participation');
-
-      const request = {
-        auth: {
-          credentials: {
-            userId,
-          },
-        },
-        params: {
-          userId,
-          campaignId,
-        },
-      };
-      const campaignParticipationSerializer = { serialize: sinon.stub() };
-      sinon.stub(usecases, 'getUserCampaignParticipationToCampaign');
-      usecases.getUserCampaignParticipationToCampaign.withArgs({ userId, campaignId }).resolves(campaignParticipation);
-      campaignParticipationSerializer.serialize.withArgs(campaignParticipation).returns(expectedCampaignParticipation);
-
-      // when
-      const response = await userController.getUserCampaignParticipationToCampaign(request, hFake, {
-        campaignParticipationSerializer,
-      });
-
-      // then
-      expect(response).to.equal(expectedCampaignParticipation);
-    });
-  });
-
   describe('#addPixAuthenticationMethodByEmail', function () {
     it('should return the user with the new pix authentication method by user email', async function () {
       // given
